# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Node MySQL API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "paths": {
    "/accounts/authenticate": {
      "post": {
        "tags": ["Accounts"],
        "summary": "Authenticate user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "user@example.com" },
                  "password": { "type": "string", "example": "password123" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Successful authentication" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/accounts/register": {
      "post": {
        "tags": ["Accounts"],
        "summary": "Register new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" },
                  "confirmPassword": { "type": "string" }
                },
                "required": ["email", "password", "confirmPassword"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Registration successful" }
        }
      }
    },
    "/accounts/verify-email": {
      "post": {
        "tags": ["Accounts"],
        "summary": "Verify email address",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": { "type": "string" }
                },
                "required": ["token"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Email verified" }
        }
      }
    },
    "/accounts/forgot-password": {
      "post": {
        "tags": ["Accounts"],
        "summary": "Request password reset",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" }
                },
                "required": ["email"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Password reset email sent" }
        }
      }
    },
    "/accounts/reset-password": {
      "post": {
        "tags": ["Accounts"],
        "summary": "Reset password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": { "type": "string" },
                  "password": { "type": "string" },
                  "confirmPassword": { "type": "string" }
                },
                "required": ["token", "password", "confirmPassword"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Password reset successful" }
        }
      }
    },
    "/accounts/refresh-token": {
      "post": {
        "tags": ["Accounts"],
        "summary": "Refresh access token",
        "responses": {
          "200": { "description": "Token refreshed" }
        }
      }
    },
    "/accounts/revoke-token": {
      "post": {
        "tags": ["Accounts"],
        "summary": "Revoke refresh token",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Token revoked" }
        }
      }
    },
    "/accounts": {
      "get": {
        "tags": ["Accounts"],
        "summary": "Get all accounts",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "List of accounts" }
        }
      }
    },
    "/accounts/{id}": {
      "get": {
        "tags": ["Accounts"],
        "summary": "Get account by ID",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Account details" }
        }
      },
      "put": {
        "tags": ["Accounts"],
        "summary": "Update account",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": { "type": "string" },
                  "lastName": { "type": "string" },
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Account updated" }
        }
      },
      "delete": {
        "tags": ["Accounts"],
        "summary": "Delete account",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Account deleted" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}